refactor(gdn): align Triton/Python GDN with SGL main (V-first state + fp32 beta)
- gdn_gating: fp32 beta output (was bf16, lossy sigmoid truncation;
  matches SGL fused_gdn_gating_prefill and FlashQLA).
- chunk_delta_h / chunk_o / fused_recurrent: V-first b_h ([BV, BK],
  byte v*K + k); h alloc (..., V, K); decode sum axis flipped to inner.
- chunk.py: docstring final_state -> [N, H, V, K]; drop the stale
  K != V raise.
- chunk_delta_h_gluon: dispatch gated off until the Gluon path is
  ported (MI300X never dispatched it; MI355X falls back to Triton).
- New test_flydsl_chunk_gdn_cache_store: aligned/tail/varlen + decode
  round-trip + output_final_state guard, 5/5 pass.

Perf (rocprofv3, Qwen3.5-9B TP1 (16,32,128,128)):
  decode -45% (B=1: 41.5 -> 22.3 us; fused_recurrent_fwd_kernel -64%).
  FlyDSL prefill within run-to-run noise.

+++ rtp_llm/models_py/triton_kernels/fla/gdn_gating.py
@@ -75,7 +75,10 @@ def fused_gdn_gating(
 
     grid = (batch, seq_len, triton.cdiv(num_heads, 8))
     g = torch.empty(1, batch, num_heads, dtype=torch.float32, device=a.device)
-    beta_output = torch.empty(1, batch, num_heads, dtype=b.dtype, device=b.device)
+    # Match SGL's fused_gdn_gating_prefill: keep beta in fp32 to avoid lossy
+    # sigmoid -> bf16 truncation before the chunk / recurrent kernels promote
+    # it back to fp32 internally.
+    beta_output = torch.empty(1, batch, num_heads, dtype=torch.float32, device=b.device)
     fused_gdn_gating_kernel[grid](
         g,
         beta_output,
